backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
cohere
qdrant-client
beautifulsoup4
requests
openai
psycopg[binary]
markdown